Implemented testing framework
I cleaned up the code and implemented a testing framework which can read accurate values from a CSV file. I'm getting 97% accuracy in a second run (weights loaded from the first run) on a 135 page PDF now. There is still no "apply this model to the rest of the document feature.