Implemented application
This is an early version of the functionality that was the purpose of the entire project: to apply a previously trained model to a PDF file so it does the work for you. You train a model on your file, it dumps the weights, and then you run the apply_model. I'm seeing 100% accurate predictions for all block types except quotes. Although this requires almost as much training data that it doesn't really make the task easier. Machine learning is just not good at specificity. Everything is a bit messy and in a preliminary state.